rename package to alloydbconnector
