+++ README.md
@@ -106,10 +106,10 @@ AlloyDB private IP address.
 Please see [Configuring AlloyDB Connectivity][alloydb-connectivity] for more details.
 
 There are two ways to import this AlloyDB connector package:
-1. `import google.cloud.alloydb_connector`
+1. `import google.cloud.alloydbconnector`
 2. `import google.cloud.alloydb.connector`
 
-The preferred way is `import google.cloud.alloydb_connector` to avoid namespace
+The preferred way is `import google.cloud.alloydbconnector` to avoid namespace
 collisions with the [google-cloud-alloydb][alloydb-py-lib] package. Meanwhile,
 `import google.cloud.alloydb.connector` will continue to work forever.
 
@@ -137,7 +137,7 @@ like `user`, `password` and `db` etc.
 To use this connector with SQLAlchemy, use the `creator` argument for `sqlalchemy.create_engine`:
 
 ```python
-from google.cloud.alloydb_connector import Connector
+from google.cloud.alloydbconnector import Connector
 import sqlalchemy
 
 # initialize Connector object
@@ -197,7 +197,7 @@ calls to `connector.close()`.
 Connector as a context manager:
 
 ```python
-from google.cloud.alloydb_connector import Connector
+from google.cloud.alloydbconnector import Connector
 import sqlalchemy
 
 # initialize Connector as context manager
@@ -248,7 +248,7 @@ currently supports the following asyncio database drivers:
 
 ```python
 import asyncpg
-from google.cloud.alloydb_connector import AsyncConnector
+from google.cloud.alloydbconnector import AsyncConnector
 
 async def main():
     # initialize AsyncConnector object for connections to AlloyDB
@@ -282,7 +282,7 @@ import asyncpg
 import sqlalchemy
 from sqlalchemy.ext.asyncio import AsyncEngine, create_async_engine
 
-from google.cloud.alloydb_connector import AsyncConnector
+from google.cloud.alloydbconnector import AsyncConnector
 
 async def init_connection_pool(connector: AsyncConnector) -> AsyncEngine:
     # The AlloyDB Python Connector can be used along with SQLAlchemy using the
@@ -335,7 +335,7 @@ need for explicit calls to `connector.close()` to cleanup resources.
 
 ```python
 import asyncpg
-from google.cloud.alloydb_connector import AsyncConnector
+from google.cloud.alloydbconnector import AsyncConnector
 
 async def main():
     # initialize AsyncConnector object for connections to AlloyDB
@@ -367,7 +367,7 @@ import asyncpg
 import sqlalchemy
 from sqlalchemy.ext.asyncio import AsyncEngine, create_async_engine
 
-from google.cloud.alloydb_connector import AsyncConnector
+from google.cloud.alloydbconnector import AsyncConnector
 
 async def init_connection_pool(connector: AsyncConnector) -> AsyncEngine:
     # The AlloyDB Python Connector can be used along with SQLAlchemy using the
@@ -462,7 +462,7 @@ and `"PSC"`.
 Example:
 
 ```python
-from google.cloud.alloydb_connector import Connector
+from google.cloud.alloydbconnector import Connector
 
 import sqlalchemy
 
@@ -498,7 +498,7 @@ Python Connector:
 import logging
 
 logging.basicConfig(format="%(asctime)s [%(levelname)s]: %(message)s")
-logger = logging.getLogger(name="google.cloud.alloydb_connector")
+logger = logging.getLogger(name="google.cloud.alloydbconnector")
 logger.setLevel(logging.DEBUG)
 ```
