+++ tests/system/test_pg8000_psc.py
@@ -17,7 +17,7 @@ import os
 
 import sqlalchemy
 
-from google.cloud.alloydb_connector import Connector
+from google.cloud.alloydbconnector import Connector
 
 
 def create_sqlalchemy_engine(
