+++ tests/unit/test_async_connector.py
@@ -23,12 +23,12 @@ from mocks import FakeConnectionInfo
 from mocks import FakeCredentials
 import pytest
 
-from google.cloud.alloydb_connector import AsyncConnector
-from google.cloud.alloydb_connector import IPTypes
-from google.cloud.alloydb_connector.client import AlloyDBClient
-from google.cloud.alloydb_connector.exceptions import ClosedConnectorError
-from google.cloud.alloydb_connector.exceptions import IPTypeNotFoundError
-from google.cloud.alloydb_connector.instance import RefreshAheadCache
+from google.cloud.alloydbconnector import AsyncConnector
+from google.cloud.alloydbconnector import IPTypes
+from google.cloud.alloydbconnector.client import AlloyDBClient
+from google.cloud.alloydbconnector.exceptions import ClosedConnectorError
+from google.cloud.alloydbconnector.exceptions import IPTypeNotFoundError
+from google.cloud.alloydbconnector.instance import RefreshAheadCache
 
 ALLOYDB_API_ENDPOINT = "alloydb.googleapis.com"
 
@@ -176,7 +176,7 @@ async def test_connect_and_close(credentials: FakeCredentials) -> None:
     """
     Test that connector.connect calls asyncpg.connect and cleans up
     """
-    with patch("google.cloud.alloydb_connector.asyncpg.connect") as connect:
+    with patch("google.cloud.alloydbconnector.asyncpg.connect") as connect:
         # patch db connection creation and return plain future
         future = asyncio.Future()
         future.set_result(True)
@@ -203,7 +203,7 @@ async def test_force_refresh(credentials: FakeCredentials) -> None:
     Test that any failed connection results in a force refresh.
     """
     with patch(
-        "google.cloud.alloydb_connector.asyncpg.connect",
+        "google.cloud.alloydbconnector.asyncpg.connect",
         side_effect=Exception("connection failed"),
     ):
         connector = AsyncConnector(credentials)
@@ -254,7 +254,7 @@ async def test_context_manager_connect_and_close(
     Test that connector.connect calls asyncpg.connect and cleans up using the
     async context manager
     """
-    with patch("google.cloud.alloydb_connector.asyncpg.connect") as connect:
+    with patch("google.cloud.alloydbconnector.asyncpg.connect") as connect:
         fake_client = FakeAlloyDBClient()
         async with AsyncConnector(credentials) as connector:
             connector._client = fake_client
