+++ tests/unit/test_instance.py
@@ -20,12 +20,12 @@ import aiohttp
 from mocks import FakeAlloyDBClient
 import pytest
 
-from google.cloud.alloydb_connector.connection_info import ConnectionInfo
-from google.cloud.alloydb_connector.exceptions import RefreshError
-from google.cloud.alloydb_connector.instance import _parse_instance_uri
-from google.cloud.alloydb_connector.instance import RefreshAheadCache
-from google.cloud.alloydb_connector.refresh_utils import _is_valid
-from google.cloud.alloydb_connector.utils import generate_keys
+from google.cloud.alloydbconnector.connection_info import ConnectionInfo
+from google.cloud.alloydbconnector.exceptions import RefreshError
+from google.cloud.alloydbconnector.instance import _parse_instance_uri
+from google.cloud.alloydbconnector.instance import RefreshAheadCache
+from google.cloud.alloydbconnector.refresh_utils import _is_valid
+from google.cloud.alloydbconnector.utils import generate_keys
 
 
 @pytest.mark.parametrize(
